refactor(treeview): #WB2-92, extract handle events function to dedicated hook
